Scripts/pyproject.toml
[project]
name = "pbrt-v4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio>=2.37.0",
    "lxml>=6.0.1",
    "openai>=1.101.0",
    "opencv-python>=4.12.0.88",
    "openexr>=3.3.5",
    "pillow>=11.3.0",
    "pyyaml>=6.0.2",
    "tiktoken>=0.11.0",
    "tenacity>=9.1.2",
]

[build-system]
requires = ["maturin>=1.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"
module-name = "transformer_typ"
features = ["pyo3/extension-module"]
